Report Generation Backend (Premium Feature)

+++ Backend/Controllers/expenseController.js
@@ -2,6 +2,10 @@ const path = require('path');
 const expenseData = require('../Model/expenseModel');
 const userDB = require('../Model/userModel');
 const sequelize = require('../db');
+const yearlyReportDb = require('../Model/yearlyReaportModel');
+const XLSX = require('xlsx');
+const { uploadToS3 } = require('../Services/S3Services');
+const UrlDb = require('../Model/fileDownloadUrlModel');
 
 exports.mainHome = (req, res) => {
     res.sendFile(path.join(__dirname, '..', '..', 'Frontend', 'Views', 'homeAfterLogin.html'))
@@ -18,18 +22,32 @@ exports.addExpense = async (req, res) => {
     const expenseAmount = parseInt(body.ExpenseAmount);
     const description = body.ExpenseDesc;
     const expenseType = body.ExpenseType;
+    const currentDate = new Date();
+    const currentMonth = currentDate.getMonth() + 1;
+    const currentYear = currentDate.getFullYear();
+    const formattedDate = `${currentMonth.toString().padStart(2, '0')}-${currentYear}`;
     try {
         const result = await userDB.findByPk(id, { attributes: ['totalExpense'], transaction: t });
+        const Yearlyresult = await yearlyReportDb.findAll({ where: { userDatumId: id, year: formattedDate.toString() }, attributes: ['TotalExpense', 'year'], transaction: t });
         const totalExpense = parseInt(result.totalExpense);
-
+        let MonthlyTotalExpense = 0;
+        if (Yearlyresult.length > 0) {
+            MonthlyTotalExpense = parseInt(Yearlyresult[0].TotalExpense);
+        }
         await expenseData.create({
             expenseAmount: expenseAmount,
             description: description,
             expenseType: expenseType,
             userDatumId: id
-
         }, { transaction: t });
-
+        if (Yearlyresult && Yearlyresult.length > 0) {
+            if (Yearlyresult[0].year == formattedDate) {
+                await yearlyReportDb.update({ TotalExpense: MonthlyTotalExpense + expenseAmount }, { where: { userDatumId: id, year: formattedDate }, transaction: t });
+            }
+        }
+        else {
+            await yearlyReportDb.create({ year: formattedDate, TotalExpense: expenseAmount, userDatumId: id }, { transaction: t })
+        }
         await userDB.update({ totalExpense: totalExpense + expenseAmount }, { where: { id: id }, transaction: t });
 
         await t.commit();
@@ -60,12 +78,26 @@ exports.deleteExpenseData = async (req, res) => {
     const t = await sequelize.transaction();
     try {
         const id = req.body.id;
-        const expenseAmount = parseInt(req.body.expenseAmount);
         const userid = req.user.id;
+        const expenseAmount = parseInt(req.body.ExpenseAmount);
+        console.log(expenseAmount)
+        const moneyDataRecord = await expenseData.findOne({ where: { id: id, userDatumId: userid } });
+        const { updatedAt } = moneyDataRecord;
+        const currentMonth = updatedAt.getMonth() + 1;
+        const currentYear = updatedAt.getFullYear();
+        const formattedDate = `${currentMonth.toString().padStart(2, '0')}-${currentYear}`;
+        const yearlyResult = await yearlyReportDb.findOne({ where: { userDatumId: userid, year: formattedDate }, transaction: t });
+
+        if (!yearlyResult) {
+            throw new Error('Yearly report not found for the specified month and year.');
+        }
+        const YearlytotalExpense = parseInt(yearlyResult.TotalExpense);
         const result = await userDB.findByPk(userid, { attributes: ['totalExpense'], transaction: t });
         const totalExpense = parseInt(result.totalExpense);
+        console.log(totalExpense)
 
         await expenseData.destroy({ where: { id: id, userDatumId: userid }, transaction: t });
+        await yearlyReportDb.update({ TotalExpense: YearlytotalExpense - expenseAmount }, { where: { userDatumId: userid, year: formattedDate }, transaction: t });
         await userDB.update({ totalExpense: totalExpense - expenseAmount }, { where: { id: userid }, transaction: t })
 
         await t.commit();
@@ -80,26 +112,39 @@ exports.deleteExpenseData = async (req, res) => {
 
 exports.updateExpense = async (req, res) => {
     const t = await sequelize.transaction();
-    const body = req.body;
-    const id = body.id;
-    const userid = req.user.id;
-    const newExpenseAmount = parseInt(body.data.ExpenseAmount);
-    const newDescription = body.data.ExpenseDesc;
-    const newExpenseType = body.data.ExpenseType;
 
     try {
+        const body = req.body;
+        const id = body.id;
+        const userid = req.user.id;
+        const newExpenseAmount = parseInt(body.data.ExpenseAmount);
+        const newDescription = body.data.ExpenseDesc;
+        const newExpenseType = body.data.ExpenseType;
+
         const ExpenseData = await expenseData.findOne({ where: { id: id, userDatumId: userid }, transaction: t });
 
-        const oldExpenseAmount = ExpenseData.expenseAmount;
-        const expenseAmountDifference = newExpenseAmount - oldExpenseAmount;
+        const oldExpenseAmount = parseInt(ExpenseData.expenseAmount);
+        const updatedAt = ExpenseData.updatedAt;
+        const currentMonth = updatedAt.getMonth() + 1;
+        const currentYear = updatedAt.getFullYear();
+        const formattedDate = `${currentMonth.toString().padStart(2, '0')}-${currentYear}`;
+
         ExpenseData.expenseAmount = newExpenseAmount;
         ExpenseData.description = newDescription;
         ExpenseData.expenseType = newExpenseType;
 
         await ExpenseData.save({ transaction: t });
+
         const result = await userDB.findByPk(userid, { attributes: ['totalExpense'], transaction: t });
         const totalExpense = parseInt(result.totalExpense);
-        await userDB.update({ totalExpense: totalExpense + expenseAmountDifference }, { where: { id: userid }, transaction: t });
+        const expenseAmountDifference = oldExpenseAmount - newExpenseAmount;
+
+        const yearlyResult = await yearlyReportDb.findOne({ where: { userDatumId: userid, year: formattedDate }, transaction: t });
+        const YearlytotalExpense = parseInt(yearlyResult.TotalExpense);
+        const TotalexpenseAmountDifference = parseInt(oldExpenseAmount - newExpenseAmount);
+
+        await yearlyReportDb.update({ TotalExpense: YearlytotalExpense - TotalexpenseAmountDifference }, { where: { userDatumId: userid, year: formattedDate }, transaction: t });
+        await userDB.update({ totalExpense: totalExpense - expenseAmountDifference }, { where: { id: userid }, transaction: t });
 
         await t.commit();
         res.status(201).json({ data: 'success' })
@@ -130,3 +175,75 @@ exports.getLeaderBoardData = async (req, res) => {
         res.status(500).json({ error: error });
     }
 };
+
+exports.getViewMonetaryPage = (req, res) => {
+    res.sendFile(path.join(__dirname, "..", '..', 'Frontend', "Views", "viewMonetoryData.html"));
+}
+
+
+exports.getYearlyExpensesData = async (req, res) => {
+    try {
+        const id = req.user.id;
+        const result = await yearlyReportDb.findAll({ where: { userDatumId: id } });
+        res.json(result);
+    } catch (err) {
+        res.status(500).json({ data: 'error' });
+        console.log(err);
+    }
+}
+
+exports.getDownloadUrl = async (req, res) => {
+    try {
+        const id = req.user.id;
+        const result = await UrlDb.findAll({ where: { userDatumId: id } });
+        res.json(result);
+    } catch (err) {
+        res.status(500).json({ data: 'error' });
+        console.log(err);
+    }
+}
+
+exports.downloadExpense = async (req, res) => {
+    const userId = req.user.id;
+    const date = new Date().toLocaleString().replace(/\//g, '-')
+    const { dailyData, weeklyData, monthlyData, yearlyData } = req.body;
+
+    const yearlyDataValues = [
+        ['Section', 'MonthYear', 'Total Expense'],
+        ...createDataArrayWithYearlySection('Yearly Data', yearlyData),
+    ]
+    const allData = [
+        ['Section', 'Date', 'ExpenseAmount', 'ExpenseType', 'Description'],
+        ...createDataArrayWithSection('Daily Data', dailyData),
+        ...createDataArrayWithSection('Weekly Data', weeklyData),
+        ...createDataArrayWithSection('Monthly Data', monthlyData),
+        ...yearlyDataValues
+    ];
+    function createDataArrayWithSection(section, data) {
+        return data.map(item => [section, item.date, item.ExpenseAmount, item.expenseType, item.description]);
+    }
+
+    function createDataArrayWithYearlySection(section, data) {
+        return data.map(item => [section, item.monthYear, item.totalExpense]);
+    }
+    const worksheet = XLSX.utils.aoa_to_sheet(allData);
+    const workbook = XLSX.utils.book_new();
+    XLSX.utils.book_append_sheet(workbook, worksheet, 'Combined Data');
+    const buffer = XLSX.write(workbook, { bookType: 'xlsx', type: 'buffer' });
+    const fileName = `expense${userId}/Report_date-${date}.xlsx`;
+    await uploadToS3('expensetracker08', buffer, fileName)
+        .then(async (fileUrl) => {
+            res.setHeader('Content-Disposition', 'attachment; filename=expense.xlsx');
+            res.setHeader('Content-Type', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet');
+            await UrlDb.create({
+                date: date,
+                fileUrl: fileUrl,
+                userDatumId: userId
+            });
+            res.status(200).json({ fileUrl, success: true });
+        })
+        .catch(error => {
+            console.error(error);
+            res.status(500).json({ error: 'Failed to upload file to S3' });
+        });
+}
